feat: add General Mail route
